Replaced ExpressionAST class with ExpressionSyntax protocol
- Added ExpressionSyntax protocol to represent an expression node
- Added Literal, Variable, FunctionCall, UnaryOperator, BinaryOperator and Parenthesis syntax
- Made ExpressionSyntax public
- Updated the expression parser

+++ Sources/PoieticCore/ExpressionParser/Lexer.swift
@@ -70,7 +70,7 @@ public class Lexer {
                 // Just accept it
             }
 
-            type = .float
+            type = .double
         }
         
         if accept("e") || accept("E") {
@@ -83,7 +83,7 @@ public class Lexer {
             while accept(\.isWholeNumber) || accept("_") {
                 // Just accept it
             }
-            type = .float
+            type = .double
         }
         
         if accept(\.isLetter) {
